+++ launch_experiment.py
@@ -1,8 +1,9 @@
-from micro_invnet import MicroConfig,MicroInvnet
+#from micro_invnet import MicroConfig,MicroInvnet
+from mnist import Config, InvNet
 import torch
 
 if __name__=="__main__":
-    config = MicroConfig()
+    config = Config()
 
     cuda_available = torch.cuda.is_available()
     device = torch.device(config.gpu if cuda_available else "cpu")
@@ -11,7 +12,7 @@ if __name__=="__main__":
         torch.cuda.set_device(device)
     print('training on:', device)
 
-    invnet = MicroInvnet(config.batch_size, config.output_path, config.data_dir,
+    invnet = InvNet(config.batch_size, config.output_path, config.data_dir,
                          config.lr, config.critic_iter, config.proj_iter,
                          config.hidden_size, device, config.lambda_gp, config.edge_fn, config.max_op)
     invnet.train(config.end_iter)
