+++ microstructure/__init__.py
@@ -1,3 +1,3 @@
 from .utils import MicrostructureDataset
 from .config import MicroConfig
-from .micro_invnet import MicroInvnet
+from .micro_invnet import InvNet
