+++ mnist/__init__.py
@@ -1,0 +1,2 @@
+from .config import Config
+from .mnist_invnet import InvNet
