+++ mnist/mnist_invnet.py
@@ -6,12 +6,12 @@ import torch.nn.functional as F
 import torchvision
 from torchvision import transforms, datasets
 
-from mnist_invnet.config import *
+from mnist.config import *
 from invnet import BaseInvNet
 from layers.dp_layer.DPLayer import DPLayer
 
 
-class GraphInvNet(BaseInvNet):
+class InvNet(BaseInvNet):
 
     def __init__(self,batch_size,output_path,data_dir,lr,critic_iters,\
                  proj_iters,hidden_size,device,lambda_gp,edge_fn,max_op,max_i=32,max_j=32,restore_mode=False):
@@ -36,25 +36,6 @@ class GraphInvNet(BaseInvNet):
         images=self.norm_data(images)
         real_lengths=self.dp_layer(images)
         return real_lengths.view(-1,1)
-
-    def sample(self,train=True):
-        if train:
-            try:
-                real_data = next(self.dataiter)
-            except:
-                self.dataiter = iter(self.train_loader)
-                real_data = self.dataiter.next()
-            if real_data[0].shape[0]<self.batch_size:
-                real_data=self.sample()
-        else:
-            try:
-                real_data = next(self.val_iter)
-            except:
-                self.val_iter = iter(self.val_loader)
-                real_data = self.val_iter.next()
-            if real_data[0].shape[0]<self.batch_size:
-                real_data=self.sample(train=False)
-        return real_data[0]
 
     def load_data(self):
         data_transform = transforms.Compose([
@@ -89,14 +70,15 @@ class GraphInvNet(BaseInvNet):
         fake_2 = torchvision.utils.make_grid(fake_2, nrow=8, padding=2)
         self.writer.add_image('G/images', fake_2, stats['iteration'])
 
-        dev_proj_err, dev_disc_cost=self.validate()
+        dev_proj_err, dev_disc_cost=self.validation()
         #Generating images for tensorboard display
-        lv=torch.tensor([600,780,960,1140]).view(-1,1).float().to(device)
+        mean,std=self.p1_mean,self.p1_std
+        lv=torch.tensor([mean-std,mean,mean+std,mean+2*std]).view(-1,1).float().to(self.device)
         with torch.no_grad():
             noisev=self.fixed_noise
             lv_v=lv
         noisev=noisev.float()
-        gen_images=self.G(noisev,lv_v).view((self.batch_size,-11,size,size))
+        gen_images=self.G(noisev,lv_v).view((4,-1,size,size))
         gen_images = self.norm_data(gen_images)
         real_images = stats['real_data']
         real_grid_images = torchvision.utils.make_grid(real_images[:4], nrow=8, padding=2)
@@ -114,7 +96,7 @@ class GraphInvNet(BaseInvNet):
         self.writer.add_hparams(self.hparams, metric_dict,global_step=stats['iteration'])
 
 if __name__=='__main__':
-    config=InvNetConfig()
+    config=Config()
 
 
     cuda_available = torch.cuda.is_available()
@@ -125,8 +107,8 @@ if __name__=='__main__':
     print('training on:',device)
     sys.stdout.flush()
 
-    invnet=GraphInvNet(config.batch_size,config.output_path,config.data_dir,
-                  config.lr,config.critic_iter,config.proj_iter,
-                  config.hidden_size,device,config.lambda_gp,config.edge_fn,config.max_op)
+    invnet=InvNet(config.batch_size, config.output_path, config.data_dir,
+                  config.lr, config.critic_iter, config.proj_iter,
+                  config.hidden_size, device, config.lambda_gp, config.edge_fn, config.max_op)
     invnet.train(10000)
     #TODO fix proj_loss reporting
